Create Function to Store Last 3 Teachers
Function writes the last 3 teachers that a student submitted a ticket from to LocalStorage

Works in testing but still needs to be tested with real tickets. (Function already put into submit function).
In the future the data will be implemented into staff dropdown box to display the last 3 teachers submitted first.